Unit test implemented
